ver 1.1

+++ main.py
@@ -3,6 +3,8 @@ This is a function that can help you organize your book pages.
 """
 
 import os
+import sys
+
 from PIL import Image, ImageEnhance
 import pytesseract
 from tqdm import tqdm
@@ -12,18 +14,27 @@ def del_not_img():
     print("We'll let you input two paths, and we'll delete all files which isn't images.")
     img_path = input("Please input your image file path:")
     out_img_path = input("Please input where to save new images:")
+
+    if os.path.exists(img_path) is False:
+        sys.exit(1)
+    if os.path.exists(out_img_path) is False:
+        os.makedirs(out_img_path)
+    print("Path verified finished!")
+
     file_name = os.listdir(img_path)
     for i in range(len(file_name)):
         if not file_name[i].endswith(".jpg"):
             print("there is some files not images, removing")
             if file_name[i] != "pages_number":
                 os.remove(os.path.join(img_path, file_name[i]))
+
     file_name = os.listdir(out_img_path)
     for i in range(len(file_name)):
         if not file_name[i].endswith(".jpg"):
             print("there is some files not images, removing")
             os.remove(os.path.join(out_img_path, file_name[i]))
-    print("verified finished!")
+
+    print("Content verified finished!")
     return img_path, out_img_path
 
 
@@ -33,7 +44,8 @@ def cut_pic(img_path, out_img_path):
     coordinates_split2 = (2480, 0, 4960, 3508)
     coordinates_page1 = (83, 3340, 197, 3427)
     coordinates_page2 = (2276, 3340, 2390, 3427)  # 2480, 4960, 3508
-    print("Loading...")
+
+    print("Being process of cutting...")
     pic_name = os.listdir(img_path)
     for i in tqdm(range(len(pic_name))):
         image = Image.open(os.path.join(img_path, pic_name[i]))
@@ -41,6 +53,7 @@ def cut_pic(img_path, out_img_path):
         image_cut.save(os.path.join(out_img_path, "left_" + pic_name[i]))
         image_cut = image.crop(coordinates_split2)
         image_cut.save(os.path.join(out_img_path, "right_" + pic_name[i]))
+
     new_pic_name = os.listdir(out_img_path)
     for i in tqdm(range(len(new_pic_name))):
         if os.path.exists(os.path.join(img_path, "pages_number")) is False:
@@ -53,6 +66,7 @@ def cut_pic(img_path, out_img_path):
         image_cut = image.crop(coordinates_page2)
         image_cut = ImageEnhance.Contrast(image_cut).enhance(3)
         image_cut.save(os.path.join(pages_path, "R-number_" + new_pic_name[i]))
+
     return new_pic_name, pages_path
 
 
@@ -61,7 +75,7 @@ def main():
     pic_name, pages_path = cut_pic(img_path, out_img_path)
     # pic_name = os.listdir(out_img_path)
     # pages_path = os.path.join(img_path, "pages_number")
-
+    print("Recognizing images' page number...")
     for i in tqdm(range(len(pic_name))):
         image = Image.open(os.path.join(pages_path, "L-number_" + pic_name[i]))
         # image = Image.open("/Users/cosz/Downloads/cc/500_bg_4009-BAA-1901A.jpg")
@@ -83,9 +97,9 @@ def main():
                     number = "error" + number
                 os.rename(os.path.join(out_img_path, pic_name[i]), os.path.join(out_img_path, number + ".jpg"))
             else:
-                print("R-number_" + pic_name[i] + "--This page is not available")
+                # print("R-number_" + pic_name[i] + "--This page is not available")
                 os.remove(os.path.join(out_img_path, pic_name[i]))
-    print("finished")
+    print("All stuff has been finished")
 
 
 if __name__ == '__main__':
